[feat] Add label and project commands - Introduced new CLI commands for managing GitHub labels and projects. - Updated version to 0.1.4.4 in pyproject.toml and __init__.py. - Refactored main.py to include new command groups for labels and projects.
