+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "fown"
-version = "0.1.4.5"
+version = "0.1.4.4"
 description = "Tiny Python CLI to automate GitHub labels and projects."
 license = "MIT"
 authors = ["bamjun <khy5116@naver.com>"]
@@ -21,9 +21,9 @@ fown = "fown.cli.main:main"
 python = "^3.12"
 click = "^8.1.7"
 PyYAML = "^6.0"
-rich = "13.8.0"
+rich = "^13.7.0"
 rich-click = "^1.7.2"
-textual = "0.52.1"
+textual = "^0.52.1"
 
 
 [tool.poetry.group.dev.dependencies]
