+++ src/fown/__init__.py
@@ -1,3 +1,3 @@
 """fown 패키지 초기화"""
 
-__version__ = "0.1.4.5"
+__version__ = "0.1.4.4"
